RAS data: add level 2 callout to ME=0 checkstop (#14)
After discussion of a field failure it was decided to add a level 2
callout at medium priority so that additional investigation is done to
determine root cause before replacing a part. The processor FRU callout
is still required by manufacturing. However, it has been moved to low
priority because it is least likely that the processor reporting this
attention actually is at fault.

Change-Id: Ic375ea6fdaae540626c9501fd0c65e21541b49ca

Signed-off-by: Zane Shelley <zshelle@us.ibm.com>